requirements.txt
beautifulsoup4==4.11.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
soupsieve==2.3.2.post1
Werkzeug==2.2.2
